fix: export useCountdown types fixes #193
This is fix for the issue where types for the hook are referenced by the shared package, which is not exported. The solve the problem now we copy the types from the shared package and add them to each package when bundling the code. This is not the best solution but it is the simplest. Once we have a good way to bundle types from monorepo this can be changed.